tools/retrieve.php: Created tool retrieve information from a list of CNPJs
This tool will receive a path for a csv file with a list of CNPJs in it, and use
it to retrieve information about those numbers. It will use the CNPJClient class
to get information from Receita's website, while decoding the captcha. The user
is informed about the tool progress by its output, that tells when a CNPJ was
succefully received, or errors happened. In the end a report containing the
success rate of the captcha decode algorithm is displayed.