activities.php: Added tool to retrieve information about activities
Companies have activities related to them. One company will have one primary
activity and zero or more secondary activities. The tool activities.php will
allow you to generate CSV files of main activities, secondary activities and a
full list of activities and their description.

The activity "00.00-0-00" is just a placeholder to companies that have no
primary activity. This situation happens when the company is closed or in
irregular fiscal situation.