First version of the tools in Python
This is a basic client and runner to allow data retrieval from
receitaws.com.br webservice. This is designed in a way that
the library will wait for a long time for a request to finish,
or try again later. Also, we are limiting the amount of parallel
requests that can be triggered at any given time.