get: added support for ReceitaWS comercial API
This will allow us to issue calls to the comercial API. That API
allows you to bypass the public (free) API queue in order to get
CNPJ data. You specify the maximum data deprecation in days that
works for you, and you will be changed only for calls that generate
new data retrievals.

This fixes #10.

Signed-off-by: Vinicius Kwiecien Ruoso <vinicius.ruoso@gmail.com>